"Integrate Huggingface's transformers library and add a new script
This commit integrates the Huggingface's transformers library to replace pytorch-transformers in the cross-encoder and bi-encoder scripts. This improves compatibility with newer models and extends the functionality range of the BERT model. Additionally, it fixes the issue about loading models with GPU/CPU device compatibility. This integration also simplifies the process of resizing the token embeddings.

A new evaluate.py script is also included in this commit. The script allows easy evaluation of test, train, and valid results, hence improving the workflow of conducting experiments.

Furthermore, we have included a tutorial to guide individuals on how to train new models, getting top-k predictions, training the cross-encoder, making necessary adjustments in the evaluate.py, and evaluating the overall work."